refactor: merge ai-agent split files into language-agent.ts and behavioral-agent.ts
